Refactoring Code - Metric and command line - creating local metrics subpackage
