+++ cometr/__init__.py
@@ -1,1 +1,2 @@
 from . import global_metrics
+from . import local_metrics
