+++ cometr/global_metrics/__init__.py
@@ -2,3 +2,4 @@ from . import MSE
 from . import SSIM
 from . import PSNR
 from . import MAE
+from .. import Metric
