+++ setup.cfg
@@ -68,7 +68,7 @@ console_scripts=
     cometr.mae=cometr.global_metrics.MAE:main
     cometr.psnr=cometr.global_metrics.PSNR:main
     cometr.ssim=cometr.global_metrics.SSIM:main
-    cometr.metrics=cometr.global_metrics.Metrics:main
+    cometr.metrics=cometr.Metric:main
 
 
 [aliases]
