+++ tests/data/output.txt
@@ -1,1 +1,1 @@
-0.167840
+7.773268
